+++ NEWS.md
@@ -1,5 +1,9 @@
 # reuseme (development version)
 
+* `solve_file_name_conflicts()` now returns the number of conflicts instead of `TRUE`, `FALSE`
+
+* `rename_files2()` was rewritten. Now uses `warn_conflicts` to determine what to do. `force` is now deprecated. Is now separated in smaller functions that are easier to test and to extend.
+
 * `mark_todo_as_complete()` will save opened documents before attempting to mark as complete.
 
 * `use_todo()` should work on all platforms now.